Fix indexing of images
